chore: moved to src, remove hard-coded MDX
