--- components/Box.tsx
@@ -1,3 +1,0 @@
-import { styled } from '../stitches.config'
-
-export const Box = styled('div', {})
