--- pages/components/text.mdx
@@ -1,28 +1,0 @@
----
-title: Text
-category: Components
----
-
-import { Text } from '@atom-learning/components'
-
-# Text
-
-`Text` is our basic text component. Use it for any non-heading text.
-
-## Sizes
-
-### `sm`
-
-<Text size="sm">
-  Anything smaller than body text. E.g. "forgot your password?" prompts
-</Text>
-
-### `md`
-
-<Text> Body text. Most text in the app will use this. </Text>
-
-### `lg`
-
-<Text size="lg"> Large, non-heading text. E.g. a drop quote in an article</Text>
-
-<PropsTable for={Text} />
