+++ src/components/Box.tsx
@@ -1,0 +1,3 @@
+import { styled } from '../../stitches.config'
+
+export const Box = styled('div', {})
