+++ src/components/Flex.tsx
@@ -1,3 +1,3 @@
-import { styled } from '../stitches.config'
+import { styled } from '../../stitches.config'
 
 export const Flex = styled('div', { display: 'flex' })
