+++ src/components/InlineCode.tsx
@@ -1,4 +1,4 @@
-import { styled } from '../stitches.config'
+import { styled } from '../../stitches.config'
 
 export const InlineCode = styled('code', {
   py: '$1',
