+++ src/components/Link.tsx
@@ -1,4 +1,4 @@
-import { styled } from '../stitches.config'
+import { styled } from '../../stitches.config'
 
 export const Link = styled('a', {
   color: '$primary600',
