+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import type { AppProps } from 'next/app'
 import * as React from 'react'
 
 import { Flex, Main, Navigation } from '../components'
-import { css } from '../stitches.config'
+import { css } from '../../stitches.config'
 import { getPages } from '../utilities'
 
 css.global({
